refactor: created listener manager file to house management of listeners
